+++ cpp/src/server/ServerConfig.h
@@ -45,9 +45,9 @@ static const char* CONFIG_METRIC_COLLECTOR = "collector";
 static const char* CONFIG_PROMETHEUS = "prometheus_config";
 static const char* CONFIG_METRIC_PROMETHEUS_PORT = "port";
 
-static const std::string CONFIG_ENGINE = "engine_config";
-static const std::string CONFIG_DCBT = "use_blas_threshold";
-static const std::string CONFIG_OMP_THREAD_NUM = "omp_thread_num";
+static const char* CONFIG_ENGINE = "engine_config";
+static const char* CONFIG_DCBT = "use_blas_threshold";
+static const char* CONFIG_OMP_THREAD_NUM = "omp_thread_num";
 
 static const char* CONFIG_RESOURCE = "resource_config";
 static const char* CONFIG_RESOURCES = "resources";
@@ -57,6 +57,8 @@ static const char* CONFIG_RESOURCE_DEVICE_ID = "device_id";
 static const char* CONFIG_RESOURCE_ENABLE_LOADER = "enable_loader";
 static const char* CONFIG_RESOURCE_ENABLE_EXECUTOR = "enable_executor";
 static const char* CONFIG_RESOURCE_CONNECTIONS = "connections";
+static const char* CONFIG_SPEED_CONNECTIONS = "speed";
+static const char* CONFIG_ENDPOINT_CONNECTIONS = "connections";
 
 
 class ServerConfig {