+++ cpp/CHANGELOG.md
@@ -19,6 +19,8 @@ Please mark all change in change log and use the ticket from JIRA.
 - MS-436 - Delete vectors failed if index created with index_type: IVF_FLAT/IVF_SQ8
 - MS-450 - server hang after run stop_server.sh
 - MS-449 - Add vectors twice success, once with ids, the other no ids
+- MS-461 - Mysql meta unittest failed
+- MS-462 - Run milvus server twices, should display error
 
 ## Improvement
 - MS-327 - Clean code for milvus
@@ -74,6 +76,9 @@ Please mark all change in change log and use the ticket from JIRA.
 - MS-442 - Merge Knowhere
 - MS-445 - Rename CopyCompleted to LoadCompleted
 - MS-451 - Update server_config.template file, set GPU compute default
+- MS-455 - Distribute tasks by minimal cost in scheduler
+- MS-460 - Put transport speed as weight when choosing neighbour to execute task
+- MS-459 - Add cache for pick function in tasktable
 
 ## New Feature
 - MS-343 - Implement ResourceMgr